ingest/requirements.txt
python-dotenv
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-qdrant
qdrant-client
pypdf